TODO: GeometryEncapsulation

+++ Exceptions/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Numerics;
 
 namespace Exceptions
 {
@@ -12,7 +13,7 @@ namespace Exceptions
         {
             Console.WriteLine("Введите целое число: ");
             int n = Convert.ToInt32(Console.ReadLine());
-            int f = 1;
+            BigInteger f = 1;
             for(int i=1; i<=n; i++)
             {
                 f *= i;
